ADD Cleaner folder structure (frontend):
Description :
	The router logic is now inside the folder Router
		* previously on app.tsx for testing
	The router must have the logged logic feature
		* depends on back feature

	The Theme is now inside folder Theme
		* previously inside app.tsx for testing
		* more readability
