+++ backend/app/transcendence/src/app.controller.ts
@@ -4,7 +4,8 @@ import { AppService } from "./app.service";
 @Controller()
 export class AppController
 {
-	constructor(private readonly appService: AppService)
+	constructor(
+		private readonly appService: AppService)
 	{
 
 	}
