+++ backend/launch_dev_install.sh
@@ -1,1 +1,1 @@
-docker run -ti --rm -v $(pwd)/app:/app -p 3000:3000 dev 
+docker run -ti --rm -v $(pwd)/app:/app -p 3000:3000 --name=backend dev_back
