+++ sandbox/front/app/react-ts/src/Theme/darkMode.tsx
@@ -1,0 +1,5 @@
+import behaviourLinkOption from "./behaviourLinkOption";
+
+export const darkMode = {
+	
+};
