+++ sandbox/front/app/react-ts/src/Theme/lightMode.tsx
@@ -1,0 +1,3 @@
+import behaviourLinkOption from "./behaviourLinkOption";
+
+export const lightMode = ;
